Module Fsml.Transition

FSM Transitions

type t = State.t * Guard.t list * Action.t list * State.t

(src,guards,actions,dst) means that the FSM will go from state src to state dst whenever all guards listed in guards evaluate to true, performing, sequentially, all actions listed in actions.

val pp : Ppx_deriving_runtime.Format.formatter -> t -> Ppx_deriving_runtime.unit

val show : t -> Ppx_deriving_runtime.string

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or

Printers

val to_string : t -> string

Parsing

val keywords : Lexing.Keywords.t

val parse : Stdlib.Genlex.token Stdlib.Stream.t -> t

val of_string : string -> t

Simulation

val is_fireable : State.t -> Expr.env -> t -> bool

is_fireable src env t returns true iff transition t is fireable when the enclosing FSM is in state state and the inputs and local variables have values recorded in environment env.