Module Fsml__Fsm

Finite State Machines

type t = {

id : string;
Name

states : Fsml.State.t list;

itrans : Fsml.State.t * Fsml.Action.t list;
Initial transition

inps : string list;
Inputs

outps : string list;
Outputs

vars : string list;
Local variables

trans : Fsml.Transition.t list;

}

The static description of a FSM

val pp : Ppx_deriving_runtime.Format.formatter -> t -> Ppx_deriving_runtime.unit

val show : t -> Ppx_deriving_runtime.string

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or

val mk_trans : string -> Fsml.Transition.t

mk_trans s builds a FSM transition from a string representation. Example: mk_trans "S1 -> S1 when c=0 with k:=1" is ("S1", [ERelop ("=", EVar "c", EInt 0)],[Assign ("k", EInt 1)], "E1"). Raises Lexing.Syntax_error (with the current lookahead token) if parsing fails.

val to_string : t -> string

to_string m writes a representation of FSM m as a string using the Yojson library.

val from_string : string -> t

from_string s returns the FSM m stored in string s using the Yojson library

val to_file : fname:string -> t -> unit

to_file f writes a representation of FSM m in file f using the Yojson library.

val from_file : fname:string -> t

from_file f returns the FSM m stored in file f using the Yojson library

type ctx = {

state : Fsml.State.t;

env : Fsml.Expr.env;

}

A context is the dynamic view of a FSM. It records its current state and, in env, the value of its inputs, outputs and local variables.

val pp_ctx : Ppx_deriving_runtime.Format.formatter -> ctx -> Ppx_deriving_runtime.unit

val show_ctx : ctx -> Ppx_deriving_runtime.string

val step : ctx -> t -> ctx

step ctx m performs one simulation step, within context ctx of FSM m. The first fireable transition is selected according to the current state and value of the inputs and local variables. The actions associated to this transition are executed and both the state and context are updated accordingly. If no fireable transition is found, the context is left unchanged.