Module Fsml__Valuation

A valuation is a collection of (name,value) associations

type name = string

val pp_name : Ppx_deriving_runtime.Format.formatter -> name -> Ppx_deriving_runtime.unit

val show_name : name -> Ppx_deriving_runtime.string

val name_to_yojson : name -> Yojson.Safe.t

val name_of_yojson : Yojson.Safe.t -> name Ppx_deriving_yojson_runtime.error_or

type value = Fsml.Expr.t

val pp_value : Ppx_deriving_runtime.Format.formatter -> value -> Ppx_deriving_runtime.unit

val show_value : value -> Ppx_deriving_runtime.string

val value_to_yojson : value -> Yojson.Safe.t

val value_of_yojson : Yojson.Safe.t -> value Ppx_deriving_yojson_runtime.error_or

type t = (name * value) list

val pp : Ppx_deriving_runtime.Format.formatter -> t -> Ppx_deriving_runtime.unit

val show : t -> Ppx_deriving_runtime.string

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or

val compare : t -> t -> int

val to_string : t -> string

exception Invalid_valuation of t

val check : name list -> t -> unit

check names vs checks whether vs is a "complete" valuation wrt. to names. i.e. whether each variable listed in names has a valuation in vs and each variable listed in vs occurs in names. Raises Invalid_valuation in case of failure.

val empty : t

exception Duplicate of name

val add : name -> value -> t -> t

val remove : name -> t -> t

val mem : name -> t -> bool

val assoc : name -> t -> value